feat: 詳細ページの処理をカスタムフック化
