+++ frontend/src/app/(protected)/cards/_components/display/view/CardDetailView.tsx
@@ -1,11 +1,8 @@
 'use client';
 
 import CardModal from '@/app/(protected)/cards/_components/modal';
-import { useModal } from '@/hooks/useModal';
-import { useCardMutations } from '@/app/(protected)/cards/_hooks/useCardMutations';
-import { useRouter } from 'next/navigation';
+import { useCardDetailView } from '@/app/(protected)/cards/_hooks/useCardDetailView';
 import { CardDetail } from '@/app/(protected)/cards/_types/index';
-import { getCardDetailBreadcrumbs } from '@/lib/utils';
 import {
   DetailContainer,
   DetailCard,
@@ -20,37 +17,13 @@ interface CardDetailViewProps {
 }
 
 export default function CardDetailView({ card }: CardDetailViewProps) {
-  const router = useRouter();
-  const { deleteCard } = useCardMutations();
-  const updateModal = useModal();
-
-  const handleDelete = () => {
-    if (!confirm('本当に削除しますか？')) {
-      return;
-    }
-
-    if (card) {
-      deleteCard(
-        { bookId: card.book_id, cardId: card.id },
-        {
-          onSuccess: () => {
-            router.push('/cards');
-          },
-        },
-      );
-    }
-  };
-
-  const breadcrumbItems = getCardDetailBreadcrumbs(card.title);
+  const { breadcrumbItems, subtitle, handleDelete, updateModal } = useCardDetailView(card);
 
   return (
     <>
       <DetailContainer breadcrumbItems={breadcrumbItems}>
         <DetailCard>
-          <DetailHeader
-            title={card.title}
-            subtitle={`書籍名: ${card.book.title}`}
-          />
+          <DetailHeader title={card.title} subtitle={subtitle} />
           <DetailDescription>{card.content}</DetailDescription>
           <DetailMetadata createdAt={card.created_at} updatedAt={card.updated_at} />
           <CardActions onEdit={updateModal.open} onDelete={handleDelete} />
