+++ frontend/src/app/(protected)/lists/_components/display/view/ListDetailView.tsx
@@ -1,7 +1,6 @@
 'use client';
 
-import { useModal } from '@/hooks/useModal';
-import { getListDetailBreadcrumbs } from '@/lib/utils';
+import { useListDetailView } from '@/app/(protected)/lists/_hooks/useListDetailView';
 import UpdateListFormModal from '@/app/(protected)/lists/_components/modal';
 import AddBookModal from '@/app/(protected)/listBooks/_components/modal/AddBookModal';
 import AddedBooks from '@/app/(protected)/lists/_components/display/AddedBooks';
@@ -12,9 +11,7 @@ import {
   DetailDescription,
   DetailMetadata,
 } from '@/components/details';
-import PublicBadge from '@/components/badge/PublicBadge';
 import ListActions from '@/app/(protected)/lists/_components/detail/ListActions';
-import { useListMutations } from '@/app/(protected)/lists/_hooks/useListMutations';
 import { ListDetail } from '@/schemas/list';
 
 interface ListDetailProps {
@@ -22,29 +19,14 @@ interface ListDetailProps {
 }
 
 export default function ListDetailView({ list }: ListDetailProps) {
-  const { deleteList } = useListMutations();
-  const updateModal = useModal();
-  const addBookModal = useModal();
-
-  const handleDelete = (id: number) => {
-    if (!confirm('本当に削除しますか？')) {
-      return;
-    }
-
-    deleteList(id);
-  };
-
-  // パンくずリストのアイテム
-  const breadcrumbItems = getListDetailBreadcrumbs(list.name);
+  const { breadcrumbItems, badges, handleDelete, updateModal, addBookModal } =
+    useListDetailView(list);
 
   return (
     <>
       <DetailContainer breadcrumbItems={breadcrumbItems}>
         <DetailCard>
-          <DetailHeader
-            title={list.name}
-            badges={<PublicBadge isPublic={list.public} />}
-          />
+          <DetailHeader title={list.name} badges={badges} />
           <DetailDescription>
             {list.description || '説明が登録されていません。'}
           </DetailDescription>
